+++ Integrations/MicrosoftTeams/MicrosoftTeams.py
@@ -11,6 +11,8 @@ import time
 from threading import Thread
 from typing import Match, Union, Optional, cast, Dict, Any, List
 import re
+from jwt.algorithms import RSAAlgorithm
+from tempfile import NamedTemporaryFile
 
 # Disable insecure warnings
 requests.packages.urllib3.disable_warnings()
@@ -430,7 +432,9 @@ def get_graph_access_token() -> str:
             return access_token
     tenant_id: str = integration_context.get('tenant_id', '')
     if not tenant_id:
-        raise ValueError('Tenant ID not found')
+        raise ValueError(
+            'Did not receive tenant ID from Microsoft Teams, verify the messaging endpoint is configured correctly.'
+        )
     url: str = f'https://login.microsoftonline.com/{tenant_id}/oauth2/v2.0/token'
     data: dict = {
         'grant_type': 'client_credentials',
@@ -525,6 +529,52 @@ def http_request(
         raise ConnectionError(error_message)
 
 
+def integration_health():
+
+    bot_framework_api_health = 'Operational'
+    graph_api_health = 'Operational'
+
+    try:
+        get_bot_access_token()
+    except ValueError as e:
+        bot_framework_api_health = f'Non operational - {str(e)}'
+
+    try:
+        get_graph_access_token()
+    except ValueError as e:
+        graph_api_health = f'Non operational - {str(e)}'
+
+    api_health_output: list = [{
+        'Bot Framework API Health': bot_framework_api_health,
+        'Graph API Health': graph_api_health
+    }]
+
+    api_health_human_readble: str = tableToMarkdown('Microsoft API Health', api_health_output)
+
+    mirrored_channels_output = list()
+    integration_context: dict = demisto.getIntegrationContext()
+    teams: list = json.loads(integration_context.get('teams', '[]'))
+    for team in teams:
+        mirrored_channels: list = team.get('mirrored_channels', [])
+        for channel in mirrored_channels:
+            mirrored_channels_output.append({
+                'Team': team.get('team_name'),
+                'Channel': channel.get('channel_name'),
+                'Investigation ID': channel.get('investigation_id')
+            })
+
+    mirrored_channels_human_readable: str
+
+    if mirrored_channels_output:
+        mirrored_channels_human_readable = tableToMarkdown(
+            'Microsoft Teams Mirrored Channels', mirrored_channels_output
+        )
+    else:
+        mirrored_channels_human_readable = 'No mirrored channels.'
+
+    demisto.results(api_health_human_readble + mirrored_channels_human_readable)
+
+
 def validate_auth_header(headers: dict) -> bool:
     """
     Validated authorization header provided in the bot activity object
@@ -537,46 +587,86 @@ def validate_auth_header(headers: dict) -> bool:
     scehma: str = parts[0]
     jwt_token: str = parts[1]
     if scehma != 'Bearer' or not jwt_token:
+        demisto.info('Authorization header validation - failed to verify schema')
         return False
+
     decoded_payload: dict = jwt.decode(jwt_token, verify=False)
     issuer: str = decoded_payload.get('iss', '')
     if issuer != 'https://api.botframework.com':
+        demisto.info('Authorization header validation - failed to verify issuer')
         return False
-    # integration_context: dict = demisto.getIntegrationContext()
-    # open_id_metadata: dict = integration_context.get('open_id_metadata', {})
-    # keys: list = open_id_metadata.get('keys', [])
-    # last_updated: int = open_id_metadata.get('last_updated', 0)
-    # if last_updated < datetime.timestamp(datetime.now() + timedelta(days=5)):
-    #     open_id_url: str = 'https://login.microsoftonline.com/common/v2.0/.well-known/openid-configuration'
-    #     response: dict = http_request('GET', open_id_url, api='bot')
-    #     jwks_uri: str = response.get('jwks_uri', '')
-    #     keys_response: dict = http_request('GET', jwks_uri, api='bot')
-    #     keys = keys_response.get('keys', [])
-    #     last_updated = datetime.timestamp(datetime.now())
-    #     open_id_metadata['keys'] = keys
-    #     open_id_metadata['last_updated'] = last_updated
-    # if not keys:
-    #     return False
-    # unverified_headers: dict = jwt.get_unverified_header(jwt_token)
-    # key_id: str = unverified_headers.get('kid', '')
-    # key_object: dict = dict()
-    # for key in keys:
-    #     if key.get('kid') == key_id:
-    #         key_object = key
-    #         break
-    # if not key_object:
-    #     return False
-    # public_key: str = RSAAlgorithm.from_jwk(json.dumps(key_object))
-    # options = {
-    #     'verify_aud': False,
-    #     'verify_exp': True
-    # }
-    # decoded_payload = jwt.decode(jwt_token, public_key, options=options)
+
+    integration_context: dict = demisto.getIntegrationContext()
+    open_id_metadata: dict = json.loads(integration_context.get('open_id_metadata', '{}'))
+    keys: list = open_id_metadata.get('keys', [])
+
+    unverified_headers: dict = jwt.get_unverified_header(jwt_token)
+    key_id: str = unverified_headers.get('kid', '')
+    key_object: dict = dict()
+
+    # Check if we got the requested key in cache
+    for key in keys:
+        if key.get('kid') == key_id:
+            key_object = key
+            break
+
+    if not key_object:
+        # Didn't find requested key in cache, getting new keys
+        try:
+            open_id_url: str = 'https://login.botframework.com/v1/.well-known/openidconfiguration'
+            response: requests.Response = requests.get(open_id_url, verify=USE_SSL)
+            if not response.ok:
+                demisto.info(f'Authorization header validation failed to fetch open ID config - {response.reason}')
+                return False
+            response_json: dict = response.json()
+            jwks_uri: str = response_json.get('jwks_uri', '')
+            keys_response: requests.Response = requests.get(jwks_uri, verify=USE_SSL)
+            if not keys_response.ok:
+                demisto.info(f'Authorization header validation failed to fetch keys - {response.reason}')
+                return False
+            keys_response_json: dict = keys_response.json()
+            keys = keys_response_json.get('keys', [])
+            open_id_metadata['keys'] = keys
+        except ValueError:
+            demisto.info('Authorization header validation - failed to parse keys response')
+            return False
+
+    if not keys:
+        # Didn't get new keys
+        demisto.info('Authorization header validation - failed to get keys')
+        return False
+
+    # Find requested key in new keys
+    for key in keys:
+        if key.get('kid') == key_id:
+            key_object = key
+            break
+
+    if not key_object:
+        # Didn't find requested key in new keys
+        demisto.info('Authorization header validation - failed to find relevant key')
+        return False
+
+    endorsements: list = key_object.get('endorsements', [])
+    if not endorsements or 'msteams' not in endorsements:
+        demisto.info('Authorization header validation - failed to verify endorsements')
+        return False
+
+    public_key: str = RSAAlgorithm.from_jwk(json.dumps(key_object))
+    options = {
+        'verify_aud': False,
+        'verify_exp': True
+    }
+    decoded_payload = jwt.decode(jwt_token, public_key, options=options)
+
     audience_claim: str = decoded_payload.get('aud', '')
     if audience_claim != demisto.params().get('bot_id'):
+        demisto.info('Authorization header validation - failed to verify audience_claim')
         return False
-    # integration_context['open_id_metadata'] = json.dumps(open_id_metadata)
-    # demisto.setIntegrationContext(integration_context)
+
+    integration_context['open_id_metadata'] = json.dumps(open_id_metadata)
+    demisto.setIntegrationContext(integration_context)
+
     return True
 
 
@@ -1330,17 +1420,52 @@ def long_running_loop():
     """
     The infinite loop which runs the mirror loop and the bot app in two different threads
     """
+
+    certificate: str = demisto.params().get('certificate', '')
+    private_key: str = demisto.params().get('key', '')
+
+    certificate_path = str()
+    private_key_path = str()
+
     try:
         port_mapping: str = PARAMS.get('longRunningPort', '')
+        port: int
         if port_mapping:
-            port: int = int(port_mapping.split(':')[0])
+            if ':' in port_mapping:
+                port = int(port_mapping.split(':')[1])
+            else:
+                port = int(port_mapping)
         else:
             raise ValueError('No port mapping was provided')
         Thread(target=channel_mirror_loop, daemon=True).start()
         demisto.info('Started channel mirror loop thread')
-        http_server = WSGIServer(('', port), APP)
-        http_server.serve_forever()
+
+        ssl_args = dict()
+
+        if certificate and private_key:
+            certificate_file = NamedTemporaryFile(delete=False)
+            certificate_path = certificate_file.name
+            certificate_file.write(bytes(certificate, 'utf-8'))
+            certificate_file.close()
+            ssl_args['certfile'] = certificate_path
+
+            private_key_file = NamedTemporaryFile(delete=False)
+            private_key_path = private_key_file.name
+            private_key_file.write(bytes(private_key, 'utf-8'))
+            private_key_file.close()
+            ssl_args['keyfile'] = private_key_path
+
+            demisto.info('Starting HTTPS Server')
+        else:
+            demisto.info('Starting HTTP Server')
+
+        server = WSGIServer(('', port), APP, **ssl_args)
+        server.serve_forever()
     except Exception as e:
+        if certificate_path:
+            os.unlink(certificate_path)
+        if private_key_path:
+            os.unlink(private_key_path)
         demisto.error(f'An error occurred in long running loop: {str(e)}')
         raise ValueError(str(e))
 
@@ -1361,7 +1486,8 @@ def main():
         'long-running-execution': long_running_loop,
         'send-notification': send_message,
         'mirror-investigation': mirror_investigation,
-        'close-channel': close_channel
+        'close-channel': close_channel,
+        'microsoft-teams-integration-health': integration_health
         # 'microsoft-teams-create-team': create_team,
         # 'microsoft-teams-send-file': send_file,
     }