+++ Integrations/AWS-Lambda/CHANGELOG.md
@@ -1,4 +1,7 @@
 ## [Unreleased]
+
+
+## [19.9.0] - 2019-09-04
 * Bugfix for Proxy/Insecure issues.
 
 ## [19.8.0] - 2019-08-06