+++ Integrations/McAfee_Advanced_Threat_Defense/CHANGELOG.md
@@ -1,4 +1,7 @@
 ## [Unreleased]
+
+
+## [19.9.0] - 2019-09-04
 Bug Fix - Improved handling of DBotScore outputs, when unsuccessfully detonating a file using the ***atd-file-upload*** command.
 
 ## [19.8.2] - 2019-08-22