+++ Integrations/Tenable_sc/CHANGELOG.md
@@ -1,4 +1,7 @@
 ## [Unreleased]
+
+
+## [19.9.0] - 2019-09-04
   - Added the **tenable-sc-get-all-scan-results** command, which retrieves all the scan results in Tenable SC.
   - Added the **Port** and **Protocol** fields to the **Hosts** output in the **get-vulnerability** command.